Update field to decimal

+++ api/datadash/datadash/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class EuroToDollar(models.Model):
     date = models.DateField()
-    priceDollar = models.IntegerField()
+    priceDollar = models.DecimalField(max_digits=3, decimal_places=2)
 
     def _str_(self):
         return self.priceDollar
